refactor: add expandable settings submenus
