+++ src/components/app-settings/ApplicationSettingsPanel.tsx
@@ -3,14 +3,17 @@ import s from "../../styles";
 import { LanguageSettingsSection } from "./LanguageSettingsSection";
 import { ShortcutSettingsSection } from "./ShortcutSettingsSection";
 import { ThemeSettingsSection } from "./ThemeSettingsSection";
+import type { SettingsPageKey } from "./types";
 
 interface ApplicationSettingsPanelProps {
+  activePage: Extract<SettingsPageKey, "general" | "theme" | "shortcuts">;
   themeMode: ThemeMode;
   systemPrefersDark: boolean;
   onThemeModeChange: (mode: ThemeMode) => void;
 }
 
 export function ApplicationSettingsPanel({
+  activePage,
   themeMode,
   systemPrefersDark,
   onThemeModeChange,
@@ -25,13 +28,17 @@ export function ApplicationSettingsPanel({
         padding: "20px",
       }}
     >
-      <LanguageSettingsSection />
-      <ThemeSettingsSection
-        themeMode={themeMode}
-        systemPrefersDark={systemPrefersDark}
-        onThemeModeChange={onThemeModeChange}
-      />
-      <ShortcutSettingsSection />
+      {activePage === "general" ? (
+        <LanguageSettingsSection />
+      ) : activePage === "theme" ? (
+        <ThemeSettingsSection
+          themeMode={themeMode}
+          systemPrefersDark={systemPrefersDark}
+          onThemeModeChange={onThemeModeChange}
+        />
+      ) : (
+        <ShortcutSettingsSection />
+      )}
     </div>
   );
 }
