+++ src/components/app-settings/types.ts
@@ -1,6 +1,14 @@
 import type { SendShortcut } from "../../shortcuts";
 
-export type NavKey = "app" | "agents" | "about";
+export type NavKey = "application" | "agents" | "about";
+export type SettingsPageKey =
+  | "general"
+  | "theme"
+  | "shortcuts"
+  | "agent-paths"
+  | "claude"
+  | "codex"
+  | "about";
 
 export interface AppSettings {
   claude_path: string;
@@ -18,9 +26,12 @@ export type AgentKey = "claude" | "codex";
 export interface AppSettingsNavItem {
   key: NavKey;
   labelKey: string;
-  logo?: string;
-  filePath?: string;
-  lang?: string;
+  children?: AppSettingsNavChild[];
+}
+
+export interface AppSettingsNavChild {
+  key: SettingsPageKey;
+  labelKey: string;
 }
 
 export const APP_SETTINGS_CHANGED_EVENT = "nezha:app-settings-changed";
