+++ src/i18n.tsx
@@ -18,12 +18,14 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.general": "General",
     "appSettings.theme": "Theme",
     "appSettings.shortcuts": "Shortcuts",
+    "appSettings.application": "Application",
     "appSettings.agentSettings": "Agent Settings",
     "appSettings.about": "About",
     "appSettings.language": "Language",
     "appSettings.appLanguage": "App Language",
     "appSettings.languageHint": "Controls the language used by the Nezha interface.",
     "appSettings.agentPaths": "Agent Installation Paths",
+    "appSettings.agentPathsShort": "Paths",
     "appSettings.autoDetect": "Auto Detect",
     "appSettings.detecting": "Detecting...",
     "appSettings.refreshVersions": "Refresh Versions",
@@ -293,12 +295,14 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.general": "通用",
     "appSettings.theme": "主题",
     "appSettings.shortcuts": "快捷键",
+    "appSettings.application": "应用",
     "appSettings.agentSettings": "Agent 设置",
     "appSettings.about": "关于",
     "appSettings.language": "语言",
     "appSettings.appLanguage": "应用语言",
     "appSettings.languageHint": "控制 Nezha 界面的显示语言。",
     "appSettings.agentPaths": "智能体安装路径",
+    "appSettings.agentPathsShort": "路径",
     "appSettings.autoDetect": "自动检测",
     "appSettings.detecting": "检测中...",
     "appSettings.refreshVersions": "刷新版本",
